requirements.txt
MetaTrader5
PyQt6
pyqtgraph
matplotlib
numpy
pandas
scikit-learn
joblib
lightgbm
xgboost
shap
pydantic
pyyaml
python-dotenv
loguru
